app1/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==20.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
pi==0.1.2
psycopg2-binary==2.9.9
pycryptodome==3.19.1
PyJWT==2.8.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
